[HTP] Name the stale-libsdkl.so failure instead of leaving it to ld.lld
Two prebuilts here have a source file called libsdkl.so -- the HexKL addon's,
and the armv9 copy meson leaves in builddir/jni/arm64-v8a -- so ndk-build
derives the same install target for both and prints "overriding recipe for
target .../libsdkl.so". The later recipe wins on a clean tree, which is the
addon's and is what the probes want, but a copy left in obj/ by an earlier
build is newer than nothing and never gets replaced. Linking then picks up
whatever revision was there first, and with a beta1 library under beta2
sources the failure surfaces as undefined sdkl_npu_get_hw_info and
sdkl_cpu_i4_rm_to_i4_wh -- which reads like a missing symbol rather than a
stale file.

Android.mk now lists both paths and says what to do, and the runbook records
the symptom next to the build step. rm -rf obj libs is the fix.

Also silences a -Wformat warning in hexkl_pin_probe: MB is unsigned long
long, so total / MB does not match %zu.

+++ test/unittest/jni_htp/hexkl_pin_probe.c
@@ -404,7 +404,7 @@ static int mode_total(int domain) {
     // (64 MiB) still brackets the failure closely enough to act on.
     if ((n_held + 1) % 4 == 0) {
       char tag[32];
-      snprintf(tag, sizeof(tag), "%zuMB", total / MB);
+      snprintf(tag, sizeof(tag), "%zuMB", (size_t)(total / MB));
       if (check_run(domain, &b, tag) != 0) {
         printf("  the kernel stopped working with %.1f MB held\n",
                (double)total / (double)MB);
